Add Hgi codeGen to HgiVulkan. (HgiGL and HgiMetal received this in a previous changelist)
With this change, HgiVulkan is up and running again and can display Renderman and Embree images with color correction.

HgiVulkan's codegen is very similar to HgiGL, except for a few changes:

- In Vulkan binding indices cannot repeat between textures and buffers
  (a texture cannot have binding index 0 and a buffer have index 0, in the same set)

- HgiVulkan (like HgiMetal) uses push constants ("shader constants") for shader params.
  (HgiGL uses a uniform block)

We may still decide to share the glsl codeGen between GL and Vulkan, but it is
too early to tell if we will share them as there are more differences than the ones above in glsl for vulkan.

ColorCorrection was not describing the 'shader constants' in the pipeline and this is a requirement.

graphicsCmds was setting the incorrect shader stages when setting the push constants.

(Internal change: 2119739)