UsdShade:
Add a concept of "containers" for connectable nodes.
UsdConnectableAPI now has an IsContainer() method, with behavior
defined by the UsdConnectableAPIBehavior defined for a prim type.
UsdShadeNodeGraph and UsdShadeMaterial act as containers;
UsdShadeShader does not.

Begin deprecation cycle for UsdConnectableAPI::Is{Shader,NodeGraph}().
We now warn on first usage unless the env var
USD_SHADE_EMIT_CONNECTABLE_API_DEPRECATION_WARNING
is set to 0.  The default is 1.

Calls these these can be replaced with IsContainer(), or if
needed, the underlying equivalent:
- GetPrim().IsA<UsdShadeShader>()
- GetPrim().IsA<UsdShadeNodeGraph>()

(Internal change: 2123154)