Fix for color channel selection task not working anymore
This fix is analogous to the fix for the selection highlighting not working anymore in change 2123834, which switched from using shader params to shader constants.

Removes the parameter buffer completely and turns _CreateParameterBuffer into _UpdateParameterBuffer. The buffer contents are not put into a buffer anymore, instead they are set via SetShaderConstants on the compositor.

There was an additional bug in that the _ParameterBuffer struct was not order in the way the shader constant parameters are order, which leads to the screenSize and channel being swapped, which effectively breaks the shader code. This is fixed and the comparison operator now considers all struct members.

The code in Execute has been rearranged to make this order requirement more clear with an additional comment

(Internal change: 2127484)