[ar 2.0] Remove filesystem path operations during layer lookup
Sdf_LayerRegistry no longer performs filesystem path operations
like normalization and absolutization when looking up layers in
the layer registry. This makes the layer registry agnostic to
whatever form of identifier and resolved path is used by the
user's resolver implementation.

It is possible for many layers in the layer registry to have
the same identifier. This occurs when the identifier resolves
to different resolved paths due to the asset resolver context.
To support this, ArResolver::IsContextDependentPath was added
to allow resolver implementations to identify paths that depend
on what context is bound at resolve time. Sdf_LayerRegistry
uses this to determine which identifiers must be resolved to
uniquely identify a layer.

For example, for ArDefaultResolver absolute or dotted relative
filesystem paths (i.e. "./foo/bar.usda") are not context-dependent.
"Search paths" (i.e. "foo/bar.usda") are context-dependent,
because they may resolve to different locations depending on
the ArDefaultResolverContext bound at resolve time. This meant
that multiple layers with the same search path identifier could
exist at the same time. Prior to this change, this only happened
to work because search paths were recognized as relative paths,
causing Sdf_LayerRegistry to use the resolved paths for lookups
in the registry. The new ArResolver API makes this more explicit
and hopefully less confusing for other implementations to provide
similar behavior.

(Internal change: 2130533)