Refactor of face culling to remove fragment shader discards when possible. Face culling should be done by the hardware (glCullFace) whenever possible rather than via discards in the FS. The gl state settings are initially set in RenderPassState::Bind(), but can be overriden by GeometricShader::BindResources() depending on the whether the mesh is double sided, has a left-handed frame, has an instancer, etc.
With this change, FS discards are still used for face culling in cases in which there is an instancer, as there can be times in which instance scale or instanceTranform can result in a left-handed frame that are not known until the FS (although I think with some more refactoring it might be possible to catch those cases too).

Additionally adds UsdImagingGL tests for hardware face culling. Also adds UpdateTransform and UpdateInstancer functions to HdUnitTestDelegate and edits UsdImagingGL_UnitTestGLDrawing to take in the specific cull style as an argument, rather than a boolean.

(Internal change: 2130844)