Fix bug that caused parameters from a shader to be incorrectly carried over to the next shader that was parsed.
This was due to RmanOslParserPlugin reusing the same
RixShaderQuery for each shader it handled. Although the
documentation for RixShaderQuery/RixShaderInfo state
that clients can re-use the object by calling Open
multiple times with different shaders, there appears
to be a bug where doing so causes parameters from
the first shader to get carried over to the next one,
and so on. The fix simply creates a new RixShaderQuery
object for each shader it parses.

Fixes #1396

(Internal change: 2131039)