[HdSt,HgiGL] Changed Ptex layout texture format
Changed the storage format of Ptex layout data to use
3 * HgiFormatUInt16Vec2 texels for each face in order improve
compatibility on platforms which don't support 3-component
textures.

An earlier change to store the layout data in a 1D texture
array instead of in a texture buffer used HgiFormatInt32Vec3
as the storage format. This is convenient as it allows for
packing the data using 1 textel per face.  Unfortunately,
that storage format is not supported on all platforms.

Also, fixed related entries in the HgiGL pixel format
conversions table.

(Internal change: 2136793)