[HdSt,HioOIIO] Init target format for image data
The HioOIIO OpenImageIO plugin supports format conversion
when loading image data, so it is important to specify the
target format in the destination image spec.

This fixes HdStTextureUtils to initialize the format field of
the destination image spec and also adds defensive code
to the HioOIIO image loading code to prevent a crash when
trying to convert to an invalid format.

(Internal change: 2141522)