Allow controlling the static capacity of the SampledValueStore arrays.
This removes the hardcoded static capacity of 4, and templates
SampledValueStore to use the static capacity specified by the client
code.

Additionally, the internal _ExecuteSampledComputations() method is now
templated so that the static capacity of its value store arrays are also
controlled by the client.

Factored out some parts of the _ExecuteSampledComputations()
implementation that do not rely on the template parameter. This limits
the amount of code moved into the header (in particular, this also hides
the usage of internal classes such as Hd_ExtComputationContextInternal).