Prevent instantiation of NdrRegistry except as a base class.
Previously it had the API of a singleton but since it was the
base class of SdrRegistry (also a singleton) there could be two
NdrRegistry instances.  Using the NdrRegistry::GetInstance()
instance could lead to incorrect behavior plus discovering and
parsing shaders twice.

NdrRegistry no longer has GetInstance() and no longer makes
TfSingleton<NdrRegistry> a friend.  It was not using the
singleton instantiation macro so that didn't need to be removed.

This change also makes ~NdrRegistry() non-virtual.  We won't
ever destroy subclasses through an NdrRegistry pointer (and we
can't since it's protected) so there's no need to make it
virtual.  There are no other virtual methods.

(Internal change: 2151882)