Adding an additional plugInfo metadata mechanism for defining auto apply API schemas to allow context/site specific auto application of API schemas. Similar to how the KindRegistry allows any plugInfo to register additional kinds, now any plugInfo can contain AutoApplyAPISchemas metadata of the form:
"AutoApplyAPISchemas": {
    "<APISchemaName1>": {
        "apiSchemaAutoApplyTo": [
            "<TypedSchema1>", "<TypedSchema2>"
        ]
    },
    "<APISchemaName2>": {
        "apiSchemaAutoApplyTo": [
            "<TypedSchema1>", "<TypedSchema2>"
        ]
    }
}

The UsdSchemaRegistry will now search all plugins looking for this metadata and will additionally apply these API schemas as built-in APIs in the targeted types' prim definitions.

It's also worth noting that this new metadata gives us the ability to auto-apply an instance of a multiple apply schema to prim types where we couldn't do so before.

(Internal change: 2152119)