[usdImaging] Re-add usdImaging-level primvar filtering.
This is a partial rollback of 2149602.  That change also corrected the hdSt primvar filtering implementation, and those changes are staying in.

The original reason for CL 2149602 was that we only use usdImaging primvar filtering for hdStorm, and hdStorm also has its own primvar filtering mechanism, making the usdImaging code redundant (in terms of what gets uploaded to the GPU).

However, while hdStorm reads primvar data lazily (and only for primvars that pass the filter), the scene delegate API pulls primvar metadata early: existence at time T, interpolation, role.  With usdImaging filtering removed, we now pull all of that metadata even for primvars that hdStorm will later filter out.  This ends up being a significant cost relative to the cost of doing primvar filtering twice.  Short of changing the scene delegate API significantly, the only way to regain that time is to re-add usdImaging filtering.

Aside from rolling back CL2149602 for usdImaging (plus the render delegate flag to turn it on), the other change here is to add "mtlx" to the list of shader source types hdSt looks for.  This is necessary to correctly look up materialx metadata during usdImaging primvar filtering.

(Internal change: 2155954)