[HdSt] Fixed buffer allocation in vboSimpleMemory Manager
This was not calculating the correct buffer allocation size
for arrays of types as used for subdivision surface patch
vertices and arrays of clip planes.

The simple allocator is typically not used for these kinds
of data but this fix allows the simple allocator to be used
instead of the striped allocator for testing and debugging.

(Internal change: 2159648)