Maintain UsdResolveInfo validity across insignificant layer stack changes
UsdResolveInfo previously held on to an index into a
layer stack to indicate the layer containing an attribute
value. It now holds on to a SdfLayerHandle directly to
ensure that additions/removals of empty layers to the
layer stack don't invalidate the object. This means
UsdAttributeQuery objects (which rely on UsdResolveInfo)
will still be valid across such changes.

Fixes #1500

(Internal change: 2161341)