Fixing a specific bug that would occur when a CoordSys is bound to an Xform prim with a renderable (Rprim) descendants and a scene description change occurs that requires the both the Xform and its descendants to be resynced. Prior to the fix of this bug the CoordSys would be resynced but none of its descendant Rprims would be dirtied.
The cause was an optimization in the UsdImagingDelegate's change processing
that assumes that all hydra prims are leaf prims in the USD scene (with some
very specific exceptions). This optimization would cause the resync of the
Xform prim to assume the CoordSys was a leaf prim and stop looking for
descendants that need to be resynced, thus the Rprims not being updated.

This fix here is updates the scene delegate to handle that coordinate system
hydra prims can be dependent on USD Xforms anywhere in the scene hierarchy
and are resynced independently from other ancestor and descendant hydra prims.

(Internal change: 2161609)