[ar 2.0] Remove uses of FetchToLocalResolvedPath in RenderMan plugin
rmanArgsParser and rmanOslParser now use the ArAsset API
instead of the FetchToLocalResolvedPath function, which is
being removed in Ar 2.0. An benefit of this is that these
parsers should now support .args and .oso files embedded in
.usdz files. However, work still needs to be done to add
this support to other parts of hdPrman.

Note that .oso assets that are *not* just files on disk are
written out to temporary files. This includes .oso files
embedded in .usdz files. This is needed because the
RixShaderQuery object used to parse OSL does not currently
support reading directly from a given buffer.

(Internal change: 2165577)
(Internal change: 2165605)