Use ArAsset to read MaterialX files in usdMtlx
This allows usdMtlx to read .mtlx files from non-filesystem
locations, including .mtlx files embedded in .usdz files.
However, this does not work with .mtlx files that are
included in other .mtlx files, since the MaterialX library
handles those itself and always expects those to be files
on disk. MaterialX would need to be enhanced to support
this in the future.

(Internal change: 2169479)