This change introduces a new feature for screen-space based widening of OpenGL curve rendering in HdStorm. It is in effect during the patch rendering mode of the curve only. The widening is based on simple metrics computed in screen space. The automatic widening kicks in when the computed curve width during the tessellation stage falls below one pixel.
Note that the current code change is structured in a way that is "always on". While this mechanism works for our production needs, we do realize other users may have different workflows. So the proposed code change is only for the core feature. We are open to adding "opt-in" features such as conditional execution of the shader code.