Fix bug when reading XInclude'd MaterialX documents embedded in .usdz files on Windows.
By default the FilePath object used by MaterialX
returns a string representation using the platform's
native path separator, which is '\' on Windows. This
runs afoul of the expected package file syntax. To
fix this we use TfNormPath to ensure the separators
are in a consistent direction across platforms.

(Internal change: 2170445)