Make UsdStage::ResolveIdentifierToEditTarget consistent with composition.
This method would previously check for and reject relative
asset paths if the edit target layer was anonymous. This
was inconsistent with how composition would treat such asset
paths if they were authored in anonymous layers. The check
has been removed. In addition, the code was modified to
ensure it would never try to resolve an anonymous layer
identifier.

This method should now always return the same path
composition would use to open the layer if the given asset
path was authored in the edit target layer. It also removes
a use of the deprecated ArResolver::IsRelativePath.

(Internal change: 2172348)