Add per-prim depth bias
The depth bias is useful to address z-fighting between 2 overlapping
gprims. A common example is mesh decal. In a CAD sketch, there can be
a lot of co-planar and overlapping objects.

The depth bias is specified in different ways in different renderers.

Unity exposes 2 glPolygonOffset parameters in shader.
https://docs.unity3d.com/2021.2/Documentation/Manual/SL-Offset.html

Unreal allows to add a world offset towards camera.
https://docs.unrealengine.com/4.26/en-US/WorkingWithContent/MeshDecals/

USD allows to set the glPolygonOffset parameters in a render pass. It
could be more flexible to specify a bias for each prim.

Just like displayColor and displayOpacity, propose to add a new builtin
primvar "primvars:depthPriority". The value is not real bias so the
name is "priority" instead of "bias".

e.g.
A gprim with primvars:depthPriority=[1] will win the same gprim with
primvars:depthPriority=[0] in depth test in GL renderer.

The depth priority is passed to renderer in the same way as other builtin
primvars. Then, Storm biases the vertex toward the camera. Similar to
the "units" parameter in glPolygonOffset function. The bias is small and
only meaningful for perfectly overlapping gprims.