Make APISchemas and AutoAppliedAPISchemas provide connectability behavior on a prim type
With this change applied API schemas can register a connectableAPIBehavior
and which when applied, also provides the connectability behavior to the Typed Prim spec.

Note that If a codeless schema specifies a implementsUsdShadeConnectableAPIBehavior
plug metadata, but has no explicit Behavior registered, a default behavior will get registered
for this codeless API schema.

Behavior Composition Rules (similar to propertySpec)
- Behavior defined for a PrimType will win unless an appliedAPISchema is applied on the prim in question, at which point behavior rules are determined from the appliedAPISchema
- PrimTypes behavior is stronger than behavior defined via an auto-applied apischema

The change also fixes a bug where a non-connectable primType incorrectly and inconsistenty returned HasConnectableAPI as "true"!
- Bug was due to the fact that we register and cache null behaviors for prim types which do not define a behavior!

Note that in a subsequent change isContainer and requiresEncapsulation
will be configured using plug metadata, similar to
implementsUsdShadeConnectableAPIBehavior metadata.

(Internal change: 2175320)