[HdSt] Improved performance and appearance of patch edge drawing
Updated the edge on surface fragment shaders to consider the
distance to patch edges only for refined meshes. This simplifies
the fragment shaders a bit when drawing unrefined meshes.

Also, when emphasizing the edges along patch boundaries we now
compute the edge blending factors using the distance to the
nearest patch edge instead of the distance to the nearest triangle
edge. This improves the appearance of patch edges.

(Internal change: 2175648)