[hd] Skip merge step for dirty lists when possible.
tl;dr:
This change removes redundant work when merging dirty lists by keeping track of which collections we've processed the dirty lists for. Rather than use HdRprimCollection, we use a wrapper struct that ignores the collection name and material tag opinions for both hashing and equality tests.
The include/exclude paths can be skipped too and will be a follow-up change.

Background:
The "dirty list merge step" of the HRenderIndex::SyncAll algorithm doesn't scale well with additional render passes (typically the case when using Storm). This step determines which Rprims need to be synced and what representations (reprs) are requested per dirty Rprim. This is done per dirty list (owned by the render pass) and thus, gets expensive with every additional render pass.

The complexity and performance sensitivity of dirty list processing in Hydra has been simplified in the last few years. From honoring the collection include/exclude paths and render tags (which were owned by HdRprimCollection then) per render-pass-managed-dirty-list, we're now in a place where render tag opinions are aggregated from the Hydra tasks and the collection paths are ignored during the SyncAll algorithm.
The original intent of this work was to unify the SyncAll algorithm for B,S and R prims such that the submitted tasks determined which prims to Sync (based on the collection path and render tag filters). Unfortunately, this work was hindered by the expression of certain Storm tasks such as shadow generation and draw targets wherein the task obtained the collection information from Sprims. This issue was dubbed the "collection paradox" wherein to know which prims to sync from the tasks, we needed to sync certain Sprims prior.

A layer of caching was also added to HdRenderIndex to cache dirty Rprim id lists in order to speed up HdDirtyList::GetDirtyRprims. This was done to improve performance in applications submitting a large number of render tasks to Hydra, such as Maya wherein a render task is submitted per Shape proxy prim.

With the removal of the collection path filter, we now often end up iterating over all Rprims in the render index _multiple times_ to generate the aggregated dirty Rprim ids. This change prevents this redundant work.

(Internal change: 2176357)